Remove image of learning pyramid
I'd recommend removing the learning pyramid image from this pill. 

The ideas of different levels of learning can be useful but unfortunately the percentage symbols imply there is some research or evidence to back up the data. In recent years its become accepted this is not the case. There is nothing wrong with encouraging learners to learn by teaching each other but having this image in the resources and having coaches possibly quote this information hurts the credibility of the good work being done.

Here's an interesting (if slightly angry and dismissive) look at the history of the learning pryamid/cone in all its guises.
http://www.willatworklearning.com/2015/01/mythical-retention-data-the-corrupted-cone.html